+++ packages/backend/src/models/RepositoryModule.ts
@@ -22,6 +22,7 @@ import {
 	MiChannelFavorite,
 	MiChannelFollowing,
 	MiChannelMuting,
+	MiDanmaku,
 	MiClip,
 	MiClipFavorite,
 	MiClipNote,
@@ -436,6 +437,12 @@ const $channelMutingRepository: Provider = {
 	inject: [DI.db],
 };
 
+const $danmakusRepository: Provider = {
+	provide: DI.danmakusRepository,
+	useFactory: (db: DataSource) => db.getRepository(MiDanmaku).extend(miRepository as MiRepository<MiDanmaku>),
+	inject: [DI.db],
+};
+
 const $registryItemsRepository: Provider = {
 	provide: DI.registryItemsRepository,
 	useFactory: (db: DataSource) => db.getRepository(MiRegistryItem).extend(miRepository as MiRepository<MiRegistryItem>),
@@ -605,6 +612,7 @@ const $reversiGamesRepository: Provider = {
 		$channelFollowingsRepository,
 		$channelFavoritesRepository,
 		$channelMutingRepository,
+		$danmakusRepository,
 		$registryItemsRepository,
 		$webhooksRepository,
 		$systemWebhooksRepository,
@@ -683,6 +691,7 @@ const $reversiGamesRepository: Provider = {
 		$channelFollowingsRepository,
 		$channelFavoritesRepository,
 		$channelMutingRepository,
+		$danmakusRepository,
 		$registryItemsRepository,
 		$webhooksRepository,
 		$systemWebhooksRepository,